LayoutTests/http/tests/security/shape-image-cors-redirect-error-message-logging-4-expected.txt

CONSOLE MESSAGE: Cross-origin redirection denied by Cross-Origin Resource Sharing policy: Redirected to either a non-HTTP URL or a URL that contains credentials.
Verify the error message in console in case of CORS failing checks.
